+++ config/splat.us.stno3.yaml
@@ -62,7 +62,8 @@ segments:
       - [0x25B4, .data, e_misc]
       - [0x26D0, data]
       - [0x3180, data]
-      - [0x3870, data]
+      - [0x3858, .data, water_effects]
+      - [0x3914, data]
       - [0x39A0, data]
       - [0x3AC0, data]
       - [0x3AF0, data]
@@ -176,8 +177,8 @@ segments:
       - [0x4AFF0, c]
       - [0x4C5A4, c, e_fire_warg]
       - [0x4F58C, c]
-      - [0x50A2C, c, player_water_effect]
-      - [0x516B4, c, e_merman2]
+      - [0x50A2C, c, water_effects]
+      - [0x52D40, c, e_merman2]
       - [0x53F38, c]
       - [0x559d0, c, e_bone_scimitar]
       - [0x56264, c, e_bat]
