+++ config/splat.us.stno4.yaml
@@ -50,7 +50,8 @@ segments:
       - [0x234C, .data, e_misc]
       - [0x2404, .data, e_particles]
       - [0x2484, .data, e_room_fg]
-      - [0x2510, data]
+      - [0x2510, .data, water_effects]
+      - [0x25CC, data]
       - [0x40980, .rodata, first_c_file]
       - [0x40B04, .rodata, e_red_door]
       - [0x40B1C, .rodata, e_collect]
@@ -72,7 +73,8 @@ segments:
       - [0x54A18, c, e_particles]
       - [0x5525C, c, e_room_fg]
       - [0x55348, c, popup]
-      - [0x55678, c, unk_55678]
+      - [0x55678, c, water_effects]
+      - [0x5798C, c, unk_55678]
       - [0x5E2E0, c, prim_helpers]
       - [0x5EB10, bss]
       - [0x5EB70, .bss, create_entity]
