+++ config/symbols.us.stno4.txt
@@ -3,10 +3,14 @@ NO4_pStObjLayoutVertical = 0x80180938;
 NO4_EntityUpdates = 0x80180A0C;
 g_EInitObtainable = 0x80180B98;
 g_EInitParticle = 0x80180BA4;
+g_EInitSpawner = 0x80180BB0;
 g_EInitInteractable = 0x80180BBC;
 g_EInitCommon = 0x80180BE0;
 g_EInitDamageNum = 0x80180BEC;
 NO4_RedDoorTiles = 0x80180D2C;
+g_SplashSfx = 0x80180EE4;
+g_SurfacingSfx = 0x80180EE6;
+g_WaterXTbl = 0x80180EE8;
 NO4_PrizeDrops = 0x80181928;
 UNK_Invincibility0 = 0x801819F4;
 g_testCollEnemyLookup = 0x80181A20;
@@ -29,6 +33,12 @@ explode_lifetime = 0x801823A0;
 g_ESoulStealOrbAngles = 0x80182404;
 g_ESoulStealOrbSprt = 0x80182414;
 g_ESoulStealOrbAnim = 0x80182474;
+g_splashAspects = 0x80182510;
+g_SplashYMovement = 0x80182530;
+g_SurfacingYTbl = 0x80182550;
+g_SideWaterSpeeds = 0x8018255C;
+g_SideWaterAngles = 0x8018259C;
+g_SmallWaterDropVel = 0x801825AC;
 D_801B0934 = 0x801C0D3C;
 NO4_EntityBackgroundBlock = 0x801C0D44;
 EntityLockCamera = 0x801C0E0C;
