+++ src/st/no3/no3.h
@@ -127,8 +127,6 @@ extern u16 D_8018127C[];
 extern u16 D_801812A0[];
 extern u16 D_801812E2[];
 extern u16 D_8018131E[];
-extern u16 g_SplashSfx;
-extern u16 g_SurfacingSfx;
 extern s16 g_WaterXTbl[];
 extern Point16 D_801813DC[];
 extern Point16 D_80181468[];
@@ -187,16 +185,8 @@ extern ObjInit D_80182764[];
 extern u16 D_80183080[];
 extern u16 D_801830A0[];
 extern u8 D_8018383C[];
-extern s32 g_SmallWaterDropVel[];
-extern s32 g_SideWaterSpeeds[];
-extern s16 g_SideWaterAngles[];
-extern s16 g_splashAspects[];
-extern s32 g_SplashYMovement[];
-extern s16 g_SurfacingYTbl[];
-extern u8 g_MediumWaterSplashAnim[];
 extern const u8 g_explosion2anim[];
 extern u8 g_FallingObject2Anim[];
-extern u8 g_HighWaterSplashAnim[];
 extern Point32 g_Merman2XY[];
 extern u16 g_MermanSplashXOffset[];
 extern u8 g_HighWaterSplashParams[];
