+++ src/st/no4/e_init.c
@@ -196,7 +196,7 @@ PfnEntityUpdate OVL_EXPORT(EntityUpdates)[] = {
 EInit D_us_80180B8C = {ANIMSET_DRA(0x03), 0x00, 0x00, 0x000, 0x000};
 EInit g_EInitObtainable = {ANIMSET_DRA(0x03), 0x00, 0x00, 0x000, 0x001};
 EInit g_EInitParticle = {ANIMSET_DRA(0x03), 0x00, 0x00, 0x000, 0x002};
-EInit D_us_80180BB0 = {ANIMSET_DRA(0x00), 0x00, 0x00, 0x000, 0x004};
+EInit g_EInitSpawner = {ANIMSET_DRA(0x00), 0x00, 0x00, 0x000, 0x004};
 EInit g_EInitInteractable = {ANIMSET_DRA(0x00), 0x00, 0x00, 0x000, 0x005};
 EInit g_EInitUnkId13 = {ANIMSET_DRA(0x00), 0x00, 0x00, 0x000, 0x002};
 EInit D_us_80180BD4 = {ANIMSET_DRA(0x00), 0x00, 0x00, 0x000, 0x001};
