+++ src/st/no4/water_effects.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "no4.h"
+
+#include "../water_effects.h"
