+++ src/st/np3/water_effects.c
@@ -1,3 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "np3.h"
-#include "../e_merman2.h"
+
+#include "../water_effects.h"
